04_python/08_grpc/pyproject.toml
[project]
name = "08-grpc"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.1",
    "grpcio>=1.74.0",
    "grpcio-tools>=1.74.0",
    "requests>=2.32.4",
]
